third_party/WebKit/LayoutTests/fast/dom/shadow/all-in-shadow-tree-expected.txt

CONSOLE WARNING: ::shadow pseudo-element is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
Test for issue 432257: Matched properties cache should work for all property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(targets[0]).width is "100px"
PASS window.getComputedStyle(targets[1]).width is "100px"
PASS successfullyParsed is true

TEST COMPLETE
